backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.0.3

# Authentication and security
PyJWT==2.8.0
bcrypt==4.1.1
python-jose[cryptography]==3.3.0

# HTTP client for API calls
httpx==0.25.2

# AWS SDK for Cloudflare R2 (S3 compatible)
boto3==1.34.0
botocore==1.34.0

# File handling
python-multipart==0.0.6
aiofiles==23.2.1

# Environment variable management
python-dotenv==1.0.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-httpx==0.25.0

# Development dependencies
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Production server
gunicorn==21.2.0 